More tests.
Move the ConjoinFixtures method to the filefixture package.  It needs to sort fixture files.  (Could've exported the that sort interface too; /shrug.)
